Better default implementation for getUser method
Default implementation for `getUser` method does not use `$userProvider`. I've spent some time wondering I could not take advantage of customer user load query (documented at https://symfony.com/doc/4.4/security/user_provider.html#using-a-custom-query-to-load-the-user). This is because default implementation bypasses the usage of passed `UserProviderInterface`.

The maker does not configure the user provider in `security.yaml`, that's why it's done this way (IMO). I think we should at least disclaim that `$userProvider` is not used, or enhance maker in order to keep this default implementation.

WDYT ?